requirements.txt
langchain
langchain-core
langchain_community
pinecone
langchain_huggingface
streamlit
pypdf2
langchain_groq
python-dotenv
langchain-pinecone
torch